Added an array variable, method to handle user input, onChange listener to the AddStudent component. Added a route to the App.js compoment
Changes made to AddStudent.js:
Added the AddStudent array variable,  and the method (handleUpdateChange) that handles the user input inside the form. Next step is to connect the backend api for addStudents to reflect on the studentsView when a user clicks on Save or Delete.

Changes made to App.js:
Added a route to route the user to the addStudent component when clicking on Add Student in the navigation bar.
Added